+++ tomcat/src/main/java/org/apache/coyote/controller/StaticResourceController.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.controller;
 
 import org.apache.coyote.http11.exception.FileNotFoundException;
-import org.apache.coyote.http11.http.ContentType;
-import org.apache.coyote.http11.http.HttpHeader;
+import org.apache.coyote.http11.http.header.ContentType;
+import org.apache.coyote.http11.http.header.HttpHeader;
 import org.apache.coyote.http11.http.HttpRequest;
 import org.apache.coyote.http11.http.HttpResponse;
 import org.apache.coyote.http11.http.HttpStatus;
