+++ tomcat/src/main/java/org/apache/coyote/http11/http/HttpRequest.java
@@ -4,6 +4,13 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.Optional;
 
+import org.apache.coyote.http11.http.header.ContentType;
+import org.apache.coyote.http11.http.header.Cookie;
+import org.apache.coyote.http11.http.header.HttpHeader;
+import org.apache.coyote.http11.http.header.HttpHeaders;
+import org.apache.coyote.http11.http.session.Session;
+import org.apache.coyote.http11.http.session.SessionManager;
+
 public class HttpRequest {
 
 	private static final String START_LINE_DELIMITER = " ";
